app/requirements.txt
numpy 
matplotlib 
scipy 
pyserial
pyqt5
